Add files via upload
A Library Management System in C++ is a software application designed to automate and manage the operations of a library efficiently. The system handles various tasks such as:

1. Book Inventory Management:
The system maintains a catalog of all available books in the library. It tracks details like book titles, authors, ISBN numbers, publication dates, and genres. It also manages the addition, deletion, and updating of book records.

2. Borrowing and Returning:
The system allows users to borrow and return books. It tracks the borrowing history, including issue and return dates. The system also checks the availability of books before issuing them and updates the inventory accordingly. Overdue fines can be calculated and recorded.

By implementing object-oriented principles, the system is modular and maintainable, making it easy to extend and modify as the library's needs evolve.